+++ acc_app_optimisation/gui/rl_train_tab.py
@@ -9,10 +9,14 @@ from PyQt5 import QtCore, QtGui, QtWidgets
 from .. import envs
 from ..job_control import rl
 from . import configuration
+from .excdialog import current_exception_dialog, exception_dialog
 from .plot_manager import PlotManager
 
 if t.TYPE_CHECKING:
-    from pyjapc import PyJapc  # pylint: disable=import-error, unused-import
+    # pylint: disable=import-error, unused-import, ungrouped-imports
+    from traceback import TracebackException
+
+    from pyjapc import PyJapc
 
 LOG = getLogger(__name__)
 
@@ -29,9 +33,6 @@ class CreatingEnvDialog(QtWidgets.QDialog):
 
 
 class RlTrainTab(QtWidgets.QWidget):
-
-    errorOccurred = QtCore.pyqtSignal()
-
     def __init__(
         self, parent: t.Optional[QtWidgets.QWidget] = None, *, plot_manager: PlotManager
     ) -> None:
@@ -59,9 +60,8 @@ class RlTrainTab(QtWidgets.QWidget):
         self._train_builder.signals.reward_lists_updated.connect(
             self._plot_manager.set_reward_curve_data
         )
-        self._train_builder.signals.training_finished.connect(
-            self._on_training_finished
-        )
+        self._train_builder.signals.run_finished.connect(self._on_training_finished)
+        self._train_builder.signals.run_failed.connect(self._on_training_failed)
         # Build the GUI.
         large = QtGui.QFont()
         large.setPointSize(12)
@@ -127,7 +127,11 @@ class RlTrainTab(QtWidgets.QWidget):
             return self._train_builder.make_env()
         except:  # pylint: disable=bare-except
             LOG.error("aborted initialization", exc_info=True)
-            self.errorOccurred.emit()
+            current_exception_dialog(
+                title="RL training",
+                text="The environment could not be initialized due to an exception",
+                parent=self.window(),
+            ).show()
             return None
         finally:
             please_wait_dialog.accept()
@@ -183,7 +187,11 @@ class RlTrainTab(QtWidgets.QWidget):
             self._current_train_job = self._train_builder.build_job()
         except:  # pylint: disable=bare-except
             LOG.error("aborted initialization", exc_info=True)
-            self.errorOccurred.emit()
+            current_exception_dialog(
+                title="RL training",
+                text="The environment could not be initialized due to an exception",
+                parent=self.window(),
+            ).show()
             return
         assert self._current_train_job is not None
         self.start_button.setEnabled(False)
@@ -201,12 +209,19 @@ class RlTrainTab(QtWidgets.QWidget):
         self.stop_button.setEnabled(False)
         self._current_train_job.cancel()
 
-    def _on_training_finished(self, success: bool) -> None:
+    def _on_training_finished(self, _cancellation_completed: bool) -> None:
         self.start_button.setEnabled(True)
         self.stop_button.setEnabled(False)
         self.save_button.setEnabled(True)
-        if not success:
-            self.errorOccurred.emit()
+
+    def _on_training_failed(self, exception: TracebackException) -> None:
+        exception_dialog(
+            exception,
+            title="RL training",
+            text="The training failed due to an exception",
+            parent=self.window(),
+        ).show()
+        self._on_training_finished(False)
 
     def _on_save_clicked(self) -> None:
         dialog = QtWidgets.QFileDialog(self.window())