+++ acc_app_optimisation/job_control/base.py
@@ -1,11 +1,22 @@
+from __future__ import annotations
+
 import abc
+import contextlib
+import sys
+import traceback
 import typing as t
 
+from cernml.coi import cancellation
 from PyQt5 import QtCore
 
 if t.TYPE_CHECKING:
     # pylint: disable = unused-import
-    from cernml.coi import cancellation
+    from logging import Logger
+    from types import TracebackType
+
+
+class BenignCancelledError(cancellation.CancelledError):
+    """Cancellation error that we raise, not the optimization problem."""
 
 
 class CannotBuildJob(Exception):
@@ -30,7 +41,7 @@ class JobBuilder(abc.ABC):
     """
 
     @abc.abstractmethod
-    def build_job(self) -> "Job":
+    def build_job(self) -> Job:
         """Build a :py:class:`Job` ready for submission.
 
         Raises:
@@ -52,10 +63,68 @@ class Job(QtCore.QRunnable):
             instead.
     """
 
-    def __init__(self, token_source: "cancellation.TokenSource") -> None:
+    def __init__(self, token_source: cancellation.TokenSource) -> None:
         super().__init__()
         self._token_source = token_source
 
     def cancel(self) -> None:
         """Send a cancellation request to the job."""
         self._token_source.cancel()
+
+
+@contextlib.contextmanager
+def catching_exceptions(
+    name: str,
+    logger: Logger,
+    *,
+    token_source: cancellation.TokenSource,
+    on_success: t.Callable[[], None],
+    on_cancel: t.Callable[[], None],
+    on_exception: t.Callable[[traceback.TracebackException], None],
+) -> t.Iterator[None]:
+    """Context manager that turns exceptions into callbacks.
+
+    This is used by jobs to catch *all* exceptions and emit them via Qt
+    signals instead. This is necessary because exiting a thread via
+    exception takes down the entire application.
+
+    This also takes care of most of the token source lifecycle
+    management for us.
+
+    Args:
+        name: A string identifying the process happening in this
+            context. This appears  only in logging messages.
+        logger: The logger where the outcome (success, cancellation,
+            exception) should be logged.
+        token_source: A token source that should be reset if possible.
+        on_success: Called if the context is exited without exception.
+        on_cancel: Called if the context is exited via
+            :exc:`~cancellation.CancelledError`.
+        on_exception: Called if the context is left through *any* other
+            exception. The argument is a
+            :exc:`~traceback.TracebackException` with as much
+            information as possible without including local variables.
+    """
+    # pylint: disable = bare-except
+    try:
+        yield
+        logger.info(f"finished {name}")
+        on_success()
+        # Catch weird race conditions: If we successfully run through and a
+        # cancellation arrives _just_ after, we automatically complete it.
+        if token_source.token.cancellation_requested:
+            token_source.token.complete_cancellation()
+            token_source.reset_cancellation()
+    except BenignCancelledError:
+        logger.info(f"cancelled {name}")
+        token_source.token.complete_cancellation()
+        token_source.reset_cancellation()
+        on_cancel()
+    except cancellation.CancelledError:
+        logger.info(f"cancelled {name}")
+        if token_source.can_reset_cancellation:
+            token_source.reset_cancellation()
+        on_cancel()
+    except:
+        logger.error(f"aborted {name}", exc_info=True)
+        on_exception(traceback.TracebackException(*sys.exc_info()))