+++ acc_app_optimisation/gui/rl_train_tab.py
@@ -211,7 +211,14 @@ class RlTrainTab(QtWidgets.QWidget):
         self.stop_button.setEnabled(False)
         self._current_train_job.cancel()
 
-    def _on_training_finished(self, _cancellation_completed: bool) -> None:
+    def _on_training_finished(self, success: bool) -> None:
+        if success:
+            QtWidgets.QMessageBox(
+                QtWidgets.QMessageBox.Information,
+                "RL training",
+                "Job has terminated successfully.",
+                parent=self.window(),
+            ).show()
         self.start_button.setEnabled(True)
         self.stop_button.setEnabled(False)
         self.save_button.setEnabled(True)