+++ acc_app_optimisation/job_control/single_objective/jobs.py
@@ -290,9 +290,7 @@ class FunctionOptimizableJob(OptJob):
             LOG,
             token_source=self._token_source,
             on_success=lambda: self._signals.optimisation_finished.emit(True),
-            on_cancel=lambda: self._signals.optimisation_finished.emit(
-                not self._token_source.cancellation_requested
-            ),
+            on_cancel=lambda: self._signals.optimisation_finished.emit(False),
             on_exception=self._signals.optimisation_failed.emit,
         ):
             LOG.info(